Split github module into two and a template The template contains the github api key and some common setup and ping functions Both modules now only raise `CODE_REPOSITORY` events a new module will be added in future to download these repos Added tests for both modules, test_module_github_codesearch is the same as before whilst test_module_github_org will spoof responses for a single repo for the organization, a member of the org and a repo belonging to the org member
