--- bbot/test/test_step_2/module_tests/test_module_github.py
@@ -1,327 +1,0 @@
-from .base import ModuleTestBase
-
-
-class TestGithub(ModuleTestBase):
-    config_overrides = {"modules": {"github": {"api_key": "asdf"}}, "omit_event_types": [], "scope_report_distance": 1}
-
-    async def setup_before_prep(self, module_test):
-        module_test.httpx_mock.add_response(url="https://api.github.com/zen")
-        module_test.httpx_mock.add_response(
-            url="https://api.github.com/search/code?per_page=100&type=Code&q=blacklanternsecurity.com&page=1",
-            json={
-                "total_count": 214,
-                "incomplete_results": False,
-                "items": [
-                    {
-                        "name": "main.go",
-                        "path": "v2/cmd/cve-annotate/main.go",
-                        "sha": "4aa7c9ec68acb4c603d4b9163bf7ed42de1939fe",
-                        "url": "https://api.github.com/repositories/252813491/contents/v2/cmd/cve-annotate/main.go?ref=06f242e5fce3439b7418877676810cbf57934875",
-                        "git_url": "https://api.github.com/repositories/252813491/git/blobs/4aa7c9ec68acb4c603d4b9163bf7ed42de1939fe",
-                        "html_url": "https://github.com/projectdiscovery/nuclei/blob/06f242e5fce3439b7418877676810cbf57934875/v2/cmd/cve-annotate/main.go",
-                        "repository": {
-                            "id": 252813491,
-                            "node_id": "MDEwOlJlcG9zaXRvcnkyNTI4MTM0OTE=",
-                            "name": "nuclei",
-                            "full_name": "projectdiscovery/nuclei",
-                            "private": False,
-                            "owner": {
-                                "login": "projectdiscovery",
-                                "id": 50994705,
-                                "node_id": "MDEyOk9yZ2FuaXphdGlvbjUwOTk0NzA1",
-                                "avatar_url": "https://avatars.githubusercontent.com/u/50994705?v=4",
-                                "gravatar_id": "",
-                                "url": "https://api.github.com/users/projectdiscovery",
-                                "html_url": "https://github.com/projectdiscovery",
-                                "followers_url": "https://api.github.com/users/projectdiscovery/followers",
-                                "following_url": "https://api.github.com/users/projectdiscovery/following{/other_user}",
-                                "gists_url": "https://api.github.com/users/projectdiscovery/gists{/gist_id}",
-                                "starred_url": "https://api.github.com/users/projectdiscovery/starred{/owner}{/repo}",
-                                "subscriptions_url": "https://api.github.com/users/projectdiscovery/subscriptions",
-                                "organizations_url": "https://api.github.com/users/projectdiscovery/orgs",
-                                "repos_url": "https://api.github.com/users/projectdiscovery/repos",
-                                "events_url": "https://api.github.com/users/projectdiscovery/events{/privacy}",
-                                "received_events_url": "https://api.github.com/users/projectdiscovery/received_events",
-                                "type": "Organization",
-                                "site_admin": False,
-                            },
-                            "html_url": "https://github.com/projectdiscovery/nuclei",
-                            "description": "Fast and customizable vulnerability scanner based on simple YAML based DSL.",
-                            "fork": False,
-                            "url": "https://api.github.com/repos/projectdiscovery/nuclei",
-                            "forks_url": "https://api.github.com/repos/projectdiscovery/nuclei/forks",
-                            "keys_url": "https://api.github.com/repos/projectdiscovery/nuclei/keys{/key_id}",
-                            "collaborators_url": "https://api.github.com/repos/projectdiscovery/nuclei/collaborators{/collaborator}",
-                            "teams_url": "https://api.github.com/repos/projectdiscovery/nuclei/teams",
-                            "hooks_url": "https://api.github.com/repos/projectdiscovery/nuclei/hooks",
-                            "issue_events_url": "https://api.github.com/repos/projectdiscovery/nuclei/issues/events{/number}",
-                            "events_url": "https://api.github.com/repos/projectdiscovery/nuclei/events",
-                            "assignees_url": "https://api.github.com/repos/projectdiscovery/nuclei/assignees{/user}",
-                            "branches_url": "https://api.github.com/repos/projectdiscovery/nuclei/branches{/branch}",
-                            "tags_url": "https://api.github.com/repos/projectdiscovery/nuclei/tags",
-                            "blobs_url": "https://api.github.com/repos/projectdiscovery/nuclei/git/blobs{/sha}",
-                            "git_tags_url": "https://api.github.com/repos/projectdiscovery/nuclei/git/tags{/sha}",
-                            "git_refs_url": "https://api.github.com/repos/projectdiscovery/nuclei/git/refs{/sha}",
-                            "trees_url": "https://api.github.com/repos/projectdiscovery/nuclei/git/trees{/sha}",
-                            "statuses_url": "https://api.github.com/repos/projectdiscovery/nuclei/statuses/{sha}",
-                            "languages_url": "https://api.github.com/repos/projectdiscovery/nuclei/languages",
-                            "stargazers_url": "https://api.github.com/repos/projectdiscovery/nuclei/stargazers",
-                            "contributors_url": "https://api.github.com/repos/projectdiscovery/nuclei/contributors",
-                            "subscribers_url": "https://api.github.com/repos/projectdiscovery/nuclei/subscribers",
-                            "subscription_url": "https://api.github.com/repos/projectdiscovery/nuclei/subscription",
-                            "commits_url": "https://api.github.com/repos/projectdiscovery/nuclei/commits{/sha}",
-                            "git_commits_url": "https://api.github.com/repos/projectdiscovery/nuclei/git/commits{/sha}",
-                            "comments_url": "https://api.github.com/repos/projectdiscovery/nuclei/comments{/number}",
-                            "issue_comment_url": "https://api.github.com/repos/projectdiscovery/nuclei/issues/comments{/number}",
-                            "contents_url": "https://api.github.com/repos/projectdiscovery/nuclei/contents/{+path}",
-                            "compare_url": "https://api.github.com/repos/projectdiscovery/nuclei/compare/{base}...{head}",
-                            "merges_url": "https://api.github.com/repos/projectdiscovery/nuclei/merges",
-                            "archive_url": "https://api.github.com/repos/projectdiscovery/nuclei/{archive_format}{/ref}",
-                            "downloads_url": "https://api.github.com/repos/projectdiscovery/nuclei/downloads",
-                            "issues_url": "https://api.github.com/repos/projectdiscovery/nuclei/issues{/number}",
-                            "pulls_url": "https://api.github.com/repos/projectdiscovery/nuclei/pulls{/number}",
-                            "milestones_url": "https://api.github.com/repos/projectdiscovery/nuclei/milestones{/number}",
-                            "notifications_url": "https://api.github.com/repos/projectdiscovery/nuclei/notifications{?since,all,participating}",
-                            "labels_url": "https://api.github.com/repos/projectdiscovery/nuclei/labels{/name}",
-                            "releases_url": "https://api.github.com/repos/projectdiscovery/nuclei/releases{/id}",
-                            "deployments_url": "https://api.github.com/repos/projectdiscovery/nuclei/deployments",
-                        },
-                        "score": 1.0,
-                    }
-                ],
-            },
-        )
-        module_test.httpx_mock.add_response(
-            url="https://api.github.com/orgs/blacklanternsecurity",
-            json={
-                "login": "blacklanternsecurity",
-                "id": 25311592,
-                "node_id": "MDEyOk9yZ2FuaXphdGlvbjI1MzExNTky",
-                "url": "https://api.github.com/orgs/blacklanternsecurity",
-                "repos_url": "https://api.github.com/orgs/blacklanternsecurity/repos",
-                "events_url": "https://api.github.com/orgs/blacklanternsecurity/events",
-                "hooks_url": "https://api.github.com/orgs/blacklanternsecurity/hooks",
-                "issues_url": "https://api.github.com/orgs/blacklanternsecurity/issues",
-                "members_url": "https://api.github.com/orgs/blacklanternsecurity/members{/member}",
-                "public_members_url": "https://api.github.com/orgs/blacklanternsecurity/public_members{/member}",
-                "avatar_url": "https://avatars.githubusercontent.com/u/25311592?v=4",
-                "description": "Security Organization",
-                "name": "Black Lantern Security",
-                "company": None,
-                "blog": "www.blacklanternsecurity.com",
-                "location": "Charleston, SC",
-                "email": None,
-                "twitter_username": None,
-                "is_verified": False,
-                "has_organization_projects": True,
-                "has_repository_projects": True,
-                "public_repos": 70,
-                "public_gists": 0,
-                "followers": 415,
-                "following": 0,
-                "html_url": "https://github.com/blacklanternsecurity",
-                "created_at": "2017-01-24T00:14:46Z",
-                "updated_at": "2022-03-28T11:39:03Z",
-                "archived_at": None,
-                "type": "Organization",
-            },
-        )
-        module_test.httpx_mock.add_response(
-            url="https://api.github.com/orgs/blacklanternsecurity/repos?per_page=100&page=1",
-            json=[
-                {
-                    "id": 459780477,
-                    "node_id": "R_kgDOG2exfQ",
-                    "name": "test_keys",
-                    "full_name": "blacklanternsecurity/test_keys",
-                    "private": False,
-                    "owner": {
-                        "login": "blacklanternsecurity",
-                        "id": 79229934,
-                        "node_id": "MDEyOk9yZ2FuaXphdGlvbjc5MjI5OTM0",
-                        "avatar_url": "https://avatars.githubusercontent.com/u/79229934?v=4",
-                        "gravatar_id": "",
-                        "url": "https://api.github.com/users/blacklanternsecurity",
-                        "html_url": "https://github.com/blacklanternsecurity",
-                        "followers_url": "https://api.github.com/users/blacklanternsecurity/followers",
-                        "following_url": "https://api.github.com/users/blacklanternsecurity/following{/other_user}",
-                        "gists_url": "https://api.github.com/users/blacklanternsecurity/gists{/gist_id}",
-                        "starred_url": "https://api.github.com/users/blacklanternsecurity/starred{/owner}{/repo}",
-                        "subscriptions_url": "https://api.github.com/users/blacklanternsecurity/subscriptions",
-                        "organizations_url": "https://api.github.com/users/blacklanternsecurity/orgs",
-                        "repos_url": "https://api.github.com/users/blacklanternsecurity/repos",
-                        "events_url": "https://api.github.com/users/blacklanternsecurity/events{/privacy}",
-                        "received_events_url": "https://api.github.com/users/blacklanternsecurity/received_events",
-                        "type": "Organization",
-                        "site_admin": False,
-                    },
-                    "html_url": "https://github.com/blacklanternsecurity/test_keys",
-                    "description": None,
-                    "fork": False,
-                    "url": "https://api.github.com/repos/blacklanternsecurity/test_keys",
-                    "forks_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/forks",
-                    "keys_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/keys{/key_id}",
-                    "collaborators_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/collaborators{/collaborator}",
-                    "teams_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/teams",
-                    "hooks_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/hooks",
-                    "issue_events_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/issues/events{/number}",
-                    "events_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/events",
-                    "assignees_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/assignees{/user}",
-                    "branches_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/branches{/branch}",
-                    "tags_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/tags",
-                    "blobs_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/git/blobs{/sha}",
-                    "git_tags_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/git/tags{/sha}",
-                    "git_refs_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/git/refs{/sha}",
-                    "trees_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/git/trees{/sha}",
-                    "statuses_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/statuses/{sha}",
-                    "languages_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/languages",
-                    "stargazers_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/stargazers",
-                    "contributors_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/contributors",
-                    "subscribers_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/subscribers",
-                    "subscription_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/subscription",
-                    "commits_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/commits{/sha}",
-                    "git_commits_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/git/commits{/sha}",
-                    "comments_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/comments{/number}",
-                    "issue_comment_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/issues/comments{/number}",
-                    "contents_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/contents/{+path}",
-                    "compare_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/compare/{base}...{head}",
-                    "merges_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/merges",
-                    "archive_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/{archive_format}{/ref}",
-                    "downloads_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/downloads",
-                    "issues_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/issues{/number}",
-                    "pulls_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/pulls{/number}",
-                    "milestones_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/milestones{/number}",
-                    "notifications_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/notifications{?since,all,participating}",
-                    "labels_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/labels{/name}",
-                    "releases_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/releases{/id}",
-                    "deployments_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/deployments",
-                    "created_at": "2022-02-15T23:10:51Z",
-                    "updated_at": "2023-09-02T12:20:13Z",
-                    "pushed_at": "2023-10-19T02:56:46Z",
-                    "git_url": "git://github.com/blacklanternsecurity/test_keys.git",
-                    "ssh_url": "git@github.com:blacklanternsecurity/test_keys.git",
-                    "clone_url": "https://github.com/blacklanternsecurity/test_keys.git",
-                    "svn_url": "https://github.com/blacklanternsecurity/test_keys",
-                    "homepage": None,
-                    "size": 2,
-                    "stargazers_count": 2,
-                    "watchers_count": 2,
-                    "language": None,
-                    "has_issues": True,
-                    "has_projects": True,
-                    "has_downloads": True,
-                    "has_wiki": True,
-                    "has_pages": False,
-                    "has_discussions": False,
-                    "forks_count": 32,
-                    "mirror_url": None,
-                    "archived": False,
-                    "disabled": False,
-                    "open_issues_count": 2,
-                    "license": None,
-                    "allow_forking": True,
-                    "is_template": False,
-                    "web_commit_signoff_required": False,
-                    "topics": [],
-                    "visibility": "public",
-                    "forks": 32,
-                    "open_issues": 2,
-                    "watchers": 2,
-                    "default_branch": "main",
-                    "permissions": {"admin": False, "maintain": False, "push": False, "triage": False, "pull": True},
-                }
-            ],
-        )
-        module_test.httpx_mock.add_response(
-            url="https://api.github.com/repos/blacklanternsecurity/test_keys/contents",
-            json=[
-                {
-                    "name": "keys",
-                    "path": "keys",
-                    "sha": "866bd4d6d04bc46b6edad65a145dd4d939c7b398",
-                    "size": 2740,
-                    "url": "https://api.github.com/repos/blacklanternsecurity/test_keys/contents/keys?ref=main",
-                    "html_url": "https://github.com/blacklanternsecurity/test_keys/blob/main/keys",
-                    "git_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/git/blobs/866bd4d6d04bc46b6edad65a145dd4d939c7b398",
-                    "download_url": "https://raw.githubusercontent.com/blacklanternsecurity/test_keys/main/keys",
-                    "type": "file",
-                    "_links": {
-                        "self": "https://api.github.com/repos/blacklanternsecurity/test_keys/contents/keys?ref=main",
-                        "git": "https://api.github.com/repos/blacklanternsecurity/test_keys/git/blobs/866bd4d6d04bc46b6edad65a145dd4d939c7b398",
-                        "html": "https://github.com/blacklanternsecurity/test_keys/blob/main/keys",
-                    },
-                },
-                {
-                    "name": "scripts",
-                    "path": "scripts",
-                    "sha": "a54582de58c87f85e91a7e808c93f97b171249e4",
-                    "size": 0,
-                    "url": "https://api.github.com/repos/blacklanternsecurity/test_keys/contents/scripts?ref=main",
-                    "html_url": "https://github.com/blacklanternsecurity/test_keys/tree/main/scripts",
-                    "git_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/git/trees/a54582de58c87f85e91a7e808c93f97b171249e4",
-                    "download_url": None,
-                    "type": "dir",
-                    "_links": {
-                        "self": "https://api.github.com/repos/blacklanternsecurity/test_keys/contents/scripts?ref=main",
-                        "git": "https://api.github.com/repos/blacklanternsecurity/test_keys/git/trees/a54582de58c87f85e91a7e808c93f97b171249e4",
-                        "html": "https://github.com/blacklanternsecurity/test_keys/tree/main/scripts",
-                    },
-                },
-                {
-                    "name": "package.json",
-                    "path": "package.json",
-                    "sha": "1162ce67f0c9840b717c39ef0c5eb4b09791e56d",
-                    "size": 484,
-                    "url": "https://api.github.com/repos/blacklanternsecurity/test_keys/contents/package.json?ref=main",
-                    "html_url": "https://github.com/blacklanternsecurity/test_keys/blob/main/package.json",
-                    "git_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/git/blobs/1162ce67f0c9840b717c39ef0c5eb4b09791e56d",
-                    "download_url": "https://raw.githubusercontent.com/blacklanternsecurity/test_keys/main/package.json",
-                    "type": "file",
-                    "_links": {
-                        "self": "https://api.github.com/repos/blacklanternsecurity/test_keys/contents/package.json?ref=main",
-                        "git": "https://api.github.com/repos/blacklanternsecurity/test_keys/git/blobs/1162ce67f0c9840b717c39ef0c5eb4b09791e56d",
-                        "html": "https://github.com/blacklanternsecurity/test_keys/blob/main/package.json",
-                    },
-                },
-            ],
-        )
-        module_test.httpx_mock.add_response(
-            url="https://api.github.com/repos/blacklanternsecurity/test_keys/contents/scripts",
-            json=[
-                {
-                    "name": "new_key",
-                    "path": "scripts/new_key",
-                    "sha": "19a03f23f36992539ba51793e3f052808ffb10fc",
-                    "size": 149,
-                    "url": "https://api.github.com/repos/blacklanternsecurity/test_keys/contents/scripts/new_key?ref=main",
-                    "html_url": "https://github.com/blacklanternsecurity/test_keys/blob/main/scripts/new_key",
-                    "git_url": "https://api.github.com/repos/blacklanternsecurity/test_keys/git/blobs/19a03f23f36992539ba51793e3f052808ffb10fc",
-                    "download_url": "https://raw.githubusercontent.com/blacklanternsecurity/test_keys/main/scripts/new_key",
-                    "type": "file",
-                    "_links": {
-                        "self": "https://api.github.com/repos/blacklanternsecurity/test_keys/contents/scripts/new_key?ref=main",
-                        "git": "https://api.github.com/repos/blacklanternsecurity/test_keys/git/blobs/19a03f23f36992539ba51793e3f052808ffb10fc",
-                        "html": "https://github.com/blacklanternsecurity/test_keys/blob/main/scripts/new_key",
-                    },
-                }
-            ],
-        )
-
-    def check(self, module_test, events):
-        discovered_repos = []
-        discovered_files = []
-        for e in events:
-            if e.type == "CODE_REPOSITORY":
-                discovered_repos.append(e.data)
-            elif e.type == "URL_UNVERIFIED":
-                discovered_files.append(e.data)
-        assert discovered_repos == [
-            {"url": "https://github.com/projectdiscovery/nuclei"},
-            {"url": "https://github.com/blacklanternsecurity/test_keys"},
-        ]
-        assert discovered_files == [
-            "https://raw.githubusercontent.com/projectdiscovery/nuclei/06f242e5fce3439b7418877676810cbf57934875/v2/cmd/cve-annotate/main.go",
-            "https://raw.githubusercontent.com/blacklanternsecurity/test_keys/main/keys",
-            "https://raw.githubusercontent.com/blacklanternsecurity/test_keys/main/scripts/new_key",
-            "https://raw.githubusercontent.com/blacklanternsecurity/test_keys/main/package.json",
-        ]
